Refactoring 1
